Added enum and moved requests
